# Parse 18-digit positive integers quickly