@lemire 2020-06-25
Intriguing.