@jkeiser 2020-06-25
Apparently lower branch misses could account for at least some of the throughput increase, but I was expecting a decrease in instruction count to be the bulk of it.