@lemire 2020-06-25
Good old perf tests are back with a vengeance.