@lemire 2020-06-25
Let us hold on to any code change that touches the actual processing code until 0.4.1 (and maybe 0.4.2) is out. I'd like to iron out quality-of-life issues first.