@jkeiser 2020-06-28
I have another PR which improves a similar number of files, but *not* twitter (confusingly to me). That PR also eliminates the reparsing we do in the overflow function. Leaving this open for discussion, but that one is probably the better one to take as it results in less code.