@lemire 2020-06-30
My comments above a hilarious in light of https://github.com/simdjson/simdjson/pull/990