@jkeiser 2020-06-30
Lol yeah :) Spot on!