@lemire 2020-06-30
I suggest dropping this PR in light of the more complete one. Even if the other one wasn't faster, it is still conceptually nicer to never reparse. I have converted it to draft for this reason.