@jkeiser 2020-06-30
Agreed.